requirements.txt
attrs==19.3.0
fuzzywuzzy[speedup]==0.17.0
networkx==2.4
paho-mqtt==1.5.0
rhasspy-hermes==0.1.6
rhasspy-nlu==0.1.6
